requirements.txt
tweepy
python-dotenv
geocoder
numpy
pytrends
pandas